+++ xlwings_utils/__init__.py
@@ -2,4 +2,5 @@ from .xlwings_utils import *
 from .xlwings_utils import __version__
 from .dropbox import *
 from .nextcloud import *
+from .github import *
 from .local import *