+++ xlwings_utils/nextcloud.py
@@ -1,8 +1,11 @@
 import requests
 import xml.etree.ElementTree
 import urllib.parse
-import xlwings_utils
 import os
+import sys
+import importlib
+from pathlib import Path
+
 
 try:
     import pyodide_http
@@ -158,9 +161,9 @@ def read(path, cached=True):
 
     cached : bool
         if True (default), result will be cached
-        
+
         if False, read on each call
-        
+
     Returns
     -------
     contents of the nextcloud file : bytes
@@ -176,9 +179,9 @@ def read(path, cached=True):
     """
     if cached:
         if path in read.cache:
-            return read.cache[path]        
+            return read.cache[path]
     else:
-        read.cache={}
+        read.cache = {}
 
     _login()
     response = requests.get(_url + path, auth=_auth)
@@ -187,9 +190,12 @@ def read(path, cached=True):
     except requests.exceptions.HTTPError as e:
         raise OSError(f"file {str(path)} not found. Original message is {e}") from None
     result = response.content
-    read.cache[path]=result
+    read.cache[path] = result
     return result
-read.cache={}
+
+
+read.cache = {}
+
 
 def write(path, contents):
     """
@@ -245,3 +251,45 @@ def delete(path):
         response.raise_for_status()
     except requests.exceptions.HTTPError as e:
         raise OSError(f"file {str(path)} could not be deleted. Original message is {e}") from None
+
+
+def import_from_folder(folder_name):
+    """
+    imports a module from a nextcloud folder
+
+    Parameters
+    ----------
+    folder_name: str or Path
+        fully qualified name of the folder containing the module, e.g.
+
+        /Python/istr/istr
+
+    Returns
+    -------
+        link to module
+
+    Note
+    ----
+    If the module is already imported, no action
+    """
+    folder_name_path = Path(folder_name)
+    module_name = folder_name_path.parts[-1]
+    if module_name in sys.modules:
+        return sys.modules[module_name]
+
+    my_packages = Path("my_packages/")
+    my_packages.mkdir(parents=True, exist_ok=True)
+
+    for entry in dir(folder_name, recursive=True):
+        entry_path = Path(entry)
+        rel_path = entry_path.relative_to(folder_name_path)
+        if "__pycache__" in str(rel_path):
+            continue
+        contents = read(entry_path.as_posix())
+        (my_packages / module_name).mkdir(parents=True, exist_ok=True)
+        with open(my_packages / module_name / rel_path, "wb") as f:
+            f.write(contents)
+
+    if str(my_packages) not in sys.path:
+        sys.path = [str(my_packages)] + sys.path
+    return importlib.import_module(module_name)