+++ xlwings_utils/xlwings_utils.py
@@ -5,7 +5,7 @@
 #  /_/\_\|_|  \_/\_/  |_||_| |_| \__, ||___/ _____  \__,_| \__||_||_||___/
 #                                |___/      |_____|
 
-__version__ = "26.0.2"
+__version__ = "26.0.3"
 
 from pathlib import Path
 import sys
@@ -793,7 +793,7 @@ def timer(func):
     return wrapper
 
 
-def import_from_folder(cloud, folder_name, module_name=None):
+def import_from_folder(cloud, folder_name, repo=None, owner=None):
     """
     imports a module from a folder provided by a cloud service (dropbox)
 
@@ -813,26 +813,26 @@ def import_from_folder(cloud, folder_name, module_name=None):
 
     Note
     ----
-    If the module is already imported, no action 
+    If the module is already imported, no action
     """
     folder_name_path = Path(folder_name)
-
-    if module_name in sys.modules: 
+    module_name = folder_name_path.parts[-1]
+    if module_name in sys.modules:
         return sys.modules[module_name]
 
     my_packages = Path("/my_packages/")
     my_packages.mkdir(parents=True, exist_ok=True)
 
-    for entry in cloud.dir(folder_name, recursive=True):
+    for entry in cloud.dir(folder_name, repo=repo, owner=owner, recursive=True):
         entry_path = Path(entry)
         rel_path = entry_path.relative_to(folder_name_path)
         if "__pycache__" in str(rel_path):
             continue
-        contents = cloud.read(entry_path)
-        (my_packages/module_name).mkdir(parents=True, exist_ok=True)
+        contents = cloud.read(entry_path, repo=repo, owner=owner)
+        (my_packages / module_name).mkdir(parents=True, exist_ok=True)
         with open(my_packages / module_name / rel_path, "wb") as f:
             f.write(contents)
-            
+
     if str(my_packages) not in sys.path:
         sys.path = [str(my_packages)] + sys.path
     return importlib.import_module(module_name)